requirements.txt
flask
flask-socketio
python-dotenv
eventlet
bcrypt
requests